backend/pyproject.toml
[project]
name = "social-media-monitor"
version = "0.1.0"
description = "Social Media Monitoring FastAPI Project"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["fastapi", "uvicorn", "social media", "monitoring"]
classifiers = [
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License"
]

dependencies = [
    # FastAPI and Core
    "fastapi==0.115.6",
    "uvicorn[standard]==0.35.0",
    "python-multipart==0.0.20",
    "python-jose[cryptography]==3.5.0",

    # SuperTokens Authentication
    "supertokens-python==0.30.2",

    # Database
    "sqlalchemy==2.0.43",
    "asyncpg==0.30.0",
    "alembic==1.16.5",

    # PostgreSQL Extensions
    "pgvector==0.4.1",

    # Redis
    "redis[hiredis]==5.0.1",
    "redis-om==0.3.5",

    # OpenSearch
    "opensearch-py==3.0.0",

    # Data Processing
    "pandas==2.3.2",
    "numpy==2.3.2",
    "scikit-learn==1.7.1",
    "nltk==3.9.1",
    "spacy==3.8.7",

    # HTTP Clients for Social Media APIs
    "httpx==0.26.0",
    "tweepy==4.14.0",
    "praw==7.7.1",
    "linkedin-api==2.0.1",

    # Background Tasks & Async
    "celery[redis]==5.5.3",
    "aiofiles==23.2.1",

    # Validation & Serialization
    "pydantic==2.11.7",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",

    # Utilities
    "python-dotenv==1.0.0",
    "structlog==23.2.0",
    "click==8.1.7",
    "rich==13.7.0",

    # Testing
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "factory-boy==3.3.0",

    # Development
    "black==23.12.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0"
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
dependencies = []
